refactor: move components to pages
