+++ app/globals.css
@@ -34,7 +34,6 @@ a,
 	text-decoration: none;
 }
 
-a:hover,
 .link:hover {
 	text-decoration: underline;
 }
